requirements.txt
requests
python-dotenv
psycopg2-binary
SQLAlchemy
matplotlib
plotly 
pandas
notebook
ipykernel
PyYAML